created comparisons folder
