+++ xai_compare/comparisons/__init__.py
@@ -1,0 +1,2 @@
+from .feature_selection import FeatureSelection
+from .consistency import Consistency
